fix mistake unveiled by @darccio
