+++ contrib/net/http/http_test.go
@@ -285,39 +285,6 @@ func TestServeMuxUsesResourceNamer(t *testing.T) {
 	assert.Equal("net/http", s.Tag(ext.Component))
 }
 
-func TestServeMuxGo122Patterns(t *testing.T) {
-	mt := mocktracer.Start()
-	defer mt.Stop()
-
-	// A mux with go1.21 patterns ("/bar") and go1.22 patterns ("GET /foo")
-	mux := NewServeMux()
-	mux.HandleFunc("/bar", func(w http.ResponseWriter, r *http.Request) {})
-	mux.HandleFunc("GET /foo", func(w http.ResponseWriter, r *http.Request) {})
-
-	// Try to hit both routes
-	barW := httptest.NewRecorder()
-	mux.ServeHTTP(barW, httptest.NewRequest("GET", "/bar", nil))
-	fooW := httptest.NewRecorder()
-	mux.ServeHTTP(fooW, httptest.NewRequest("GET", "/foo", nil))
-
-	// Assert the number of spans
-	assert := assert.New(t)
-	spans := mt.FinishedSpans()
-	assert.Equal(2, len(spans))
-
-	// Check the /bar span
-	barSpan := spans[0]
-	assert.Equal(http.StatusOK, barW.Code)
-	assert.Equal("/bar", barSpan.Tag(ext.HTTPRoute))
-	assert.Equal("GET /bar", barSpan.Tag(ext.ResourceName))
-
-	// Check the /foo span
-	fooSpan := spans[1]
-	assert.Equal(http.StatusOK, fooW.Code)
-	assert.Equal("/foo", fooSpan.Tag(ext.HTTPRoute))
-	assert.Equal("GET /foo", fooSpan.Tag(ext.ResourceName))
-}
-
 func TestWrapHandlerWithResourceNameNoRace(_ *testing.T) {
 	mt := mocktracer.Start()
 	defer mt.Stop()
